QoL + Corrections.
- Незаряживаемое незаряжается. Не заряжается сверх меры.

+++ code/game/gamemodes/wizard/spellbook.dm
@@ -766,8 +766,9 @@
 	..()
 	initialize()
 
+
 /obj/item/spellbook/magic_charge_act(mob/user)
-	. |= RECHARGE_SUCCESSFUL|RECHARGE_BURNOUT
+	. = RECHARGE_SUCCESSFUL|RECHARGE_BURNOUT
 
 	to_chat(user, span_caution("Glowing red letters appear on the front cover..."))
 	to_chat(user, span_warning(pick("NICE TRY BUT NO!", \
@@ -776,7 +777,6 @@
 				"CUTE!", \
 				"YOU DIDN'T THINK IT'D BE THAT EASY, DID YOU?")))
 
-	return .
 
 /obj/item/spellbook/attackby(obj/item/I, mob/living/user, params)
 	if(user.a_intent == INTENT_HARM || skip_refunds)
@@ -1027,19 +1027,22 @@
 	uses = 1
 	desc = "This template spellbook was never meant for the eyes of man..."
 
+
 /obj/item/spellbook/oneuse/magic_charge_act(mob/user)
+	. = NONE
+
+	if(!used)
+		return
+
+	used = FALSE
 	. |= RECHARGE_SUCCESSFUL
 
 	if(prob(80))
 		visible_message(span_warning("[src] catches fire!"))
+		user.temporarily_remove_item_from_inventory(src)
 		qdel(src)
-
 		. |= RECHARGE_BURNOUT
 
-		return .
-
-	used = FALSE
-	return .
 
 /obj/item/spellbook/oneuse/New()
 	..()
